Moved additem functionality to add-item component from dashboard comp.
Created children of dashboard route

Children components are displayed in the displayComponent outlet

Switch between view and add functionality with buttons
