+++ src/app/Router/router.module.ts
@@ -6,13 +6,17 @@ import {LoginComponent} from '../login/login.component';
 import {SignupComponent} from '../signup/signup.component';
 import {DashboardComponent} from '../dashboard/dashboard.component';
 import { DisplayComponent } from '../display/display.component';
+import { AddItemComponent } from '../add-item/add-item.component';
 
 const routes: Routes = [
   {path: '', redirectTo: 'login', pathMatch: 'full'},
   {path: 'login', component: LoginComponent},
   {path: 'signUp', component: SignupComponent},
-  {path: 'dashboard', component: DashboardComponent},
-  {path: 'display', component: DisplayComponent}
+  {path: 'dashboard', component: DashboardComponent, children: [
+    {path: 'display', component: DisplayComponent, outlet: "displayComponent"},
+    {path: 'additem', component: AddItemComponent, outlet: "displayComponent"}
+  ]}
+  
 ];
