+++ src/app/add-item/add-item.component.css
@@ -1,0 +1,32 @@
+.jumbotron{
+    text-align: center;
+}
+
+
+.formFields .form-group{
+    display: flex;
+    flex-direction: column;
+    /* max-width: 80%; */
+    margin: auto;
+}
+
+.buttonGroup{
+    margin: 5px;
+    text-align: center;
+}
+
+button{
+    margin: 5px;
+}
+
+label{
+    font-weight: bold;
+}
+
+.card{
+    padding: 0;
+}
+
+.row{
+    padding: 10px;
+}
