+++ src/app/dashboard/dashboard.component.html
@@ -1,58 +1,14 @@
 <div class="container-fluid">
-  <h1 class="jumbotron shadow text-center">Purchase History</h1>
-
-  <div class="card-deck">
-    <div class="card border-primary text-center shadow-lg">
-      <h5 class="card-header border-primary">Add a new item</h5>
-      <div class="card-body">
-        <form [formGroup]='form' #myForm = 'ngForm' (ngSubmit)='onAdd(form.value)' class='formFields'>
-          <div class="form-group">
-            <label>Company Name:</label>
-            <input type="text" class='form-control' formControlName='CompanyName' placeholder="Apple lnc.">
-            <div *ngIf="form.controls['CompanyName'].touched && form.controls['CompanyName'].hasError('required')" class="alert alert-danger">Please enter the company name</div>
-          </div>
-          <div class="form-group">
-            <label>Number of shares bought:</label>
-            <input type="number" class='form-control' formControlName='NumberOfSharesBought' placeholder="1">
-            <div *ngIf="form.controls['NumberOfSharesBought'].touched && form.controls['NumberOfSharesBought'].hasError('required')" class="alert alert-danger">Please enter the number of shares bought</div>
-          </div>
-          <div class="form-group">
-            <label>Date Purchased:</label>
-            <input type="date" class='form-control' formControlName='DatePurchased'>
-            <div *ngIf="form.controls['DatePurchased'].touched && form.controls['DatePurchased'].hasError('required')" class="alert alert-danger">Please enter the purchase date</div>
-          </div>
-
-          <div class="form-group">
-            <label>Date sold:</label>
-            <input type="date" #refName class='form-control' formControlName='DateSold'>
-            <div *ngIf="form.controls.DateSold.touched && form.controls['DateSold'].hasError('required')" class="alert alert-danger">Please enter the sell date</div>
-          </div>
-          <div class="form-group">
-            <label>Number of shares sold:</label>
-            <input type="number" class='form-control' formControlName='NumberOfSharesSold' placeholder="1">
-            <div *ngIf="form.controls['NumberOfSharesSold'].touched && form.controls['NumberOfSharesSold'].hasError('required')" class="alert alert-danger">Please enter the number of shares sold</div>
-          </div>
-          <div class="form-group">
-            <label>Purchase Price:</label>
-            <input type="number" class='form-control' formControlName='PurchasePrice' placeholder="$1000">
-            <div *ngIf="form.controls['PurchasePrice'].touched && form.controls['PurchasePrice'].hasError('required')" class="alert alert-danger">Please enter the purchase price</div>
-          </div>
-          <div class="form-group">
-            <label>Sold Price:</label>
-            <input type="number" class='form-control' formControlName='SoldPrice' placeholder="$2000">
-            <div *ngIf="form.controls['SoldPrice'].touched && form.controls['SoldPrice'].hasError('required')" class="alert alert-danger">Please enter the sell price</div>
-          </div>
-          
-          <div class="buttonGroup">
-            <button [disabled]='!form.valid' type="submit" class="btn btn-primary">Add</button>
-            <button [disabled]='dataFromStore === undefined || dataFromStore.length == 0' class="btn btn-primary" (click) = 'onView()'>View</button>
-            <button [disabled]='dataFromStore === undefined || dataFromStore.length == 0' type='button' (click)='onDownload()' class="btn btn-primary">Download</button>
-          </div>
-
-        </form>
-      </div>
-    </div>
-    <!-- <app-display class="card col-md-6 shadow"></app-display> -->
-    
+  <div class="jumbotron shadow text-center">
+    <button class="btn btn-primary" (click)='onAdd()'>Add</button>
+    <button class="btn btn-primary" (click)='onView()'>View</button>
+    <h1>
+      Purchase History
+    </h1>
   </div>
+
+
+  <router-outlet name='displayComponent'></router-outlet>
+  <!-- <router-outlet name='additemComponent'></router-outlet> -->
+  <!-- <router-outlet name = 'displayComponent'></router-outlet> -->
 </div>
